Create a custom hook file for responsiveness
